component 통합
